requirements.txt
alembic==1.3.1
appnope==0.1.0
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.10.19
botocore==1.13.19
catalogue==1.0.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
coverage==5.0.2
cryptography==2.9.2
cssselect==1.1.0
cssutils==1.0.2
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.7
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
en-core-web-sm==2.2.5
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
funcy==1.14
future==0.18.2
gensim==3.8.1
html2text==2019.9.26
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.3.0
incremental==17.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
jieba3k==0.35.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kiwisolver==1.2.0
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.0.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
newspaper3k==0.2.8
nltk==3.4.5
notebook==6.0.3
numexpr==2.7.1
numpy==1.17.4
packaging==20.0
pandas==1.0.3
pandocfilters==1.4.2
parsel==1.6.0
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Protego==0.1.16
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.1
py-rouge==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pyLDAvis==2.1.2
pynliner==0.8.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyperclip==1.8.0
pyrsistent==0.15.7
pytest==5.3.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
queuelib==1.5.0
regex==2020.5.14
requests==2.22.0
requests-file==1.4.3
s3transfer==0.2.1
sacremoses==0.0.43
scipy==1.3.2
Scrapy==2.1.0
Send2Trash==1.5.0
sentencepiece==0.1.91
service-identity==18.1.0
six==1.12.0
smart-open==1.9.0
soupsieve==1.9.4
spacy==2.2.3
spacy-lookups-data==0.2.0
SQLAlchemy==1.3.11
sqlparse==0.3.1
srsly==1.0.1
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
tinysegmenter==0.3
tldextract==2.2.1
tokenizers==0.7.0
torch==1.5.0
torchvision==0.6.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
transformers==2.10.0
Twisted==20.3.0
urllib3==1.25.6
w3lib==1.22.0
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.7.0
WTForms==2.3.1
zipp==0.6.0
zope.interface==5.1.0
